requirements.txt
pyyaml>=6.0
pydantic>=1.10
pytest>=7.4
requests>=2.31
vllm>=0.11.2
transformers>=4.57.0
